RegistrationPipelineInference/requirements.txt
setuptools~=57.0.0
opencv-python==3.4.2.16
numpy~=1.19.5
packaging~=21.3
requests~=2.27.1
PyYAML~=6.0
scipy~=1.5.4
matplotlib~=3.3.4
Pillow~=8.4.0
tensorflow-gpu==1.12.0
tqdm~=4.64.1
scikit-image~=0.17.2
torch~=1.10.1
torchvision==0.11.1
dominate~=2.8.0
GPUtil~=1.4.0
